Update to bsnes v068r01 release.
This adds RTS/CTS support to the serial communications emulation.
Okay, well the PC acts as if it is always ready, because it always is
even on the real thing, but the PC-waiting-for-SNES side works.

Source only, hardware communication only works on OS X and Linux
(Windows serial communication is totally different, I don't feel like
writing a Windows version), more documentation will come later.