Update to bsnes v068r02 release.
This adds mosaic improvements to the S-PPU dot renderer. Specifically,
it eliminates the mosaic_table entirely, and performs mosaic
adjustment per pixel instead. It also moves from a mosaic countdown
for mosaic Y to a mosaic counter (incrementing).

In the process, I realized Sim Earth's map was broken, so I fixed
that. In doing so, I also fixed my old Mode7 demo that was always off-
by-one, causing different results on real hardware versus emulation.
But then I broke both Final Fantasy 5 and Air Strike Patrol effects
that use Mode7 but no mosaic.

I'm not really sure what's going on, but I think I am close. This is
the first time I can reproduce the Mode7 test ROM results without
screwing with M7Y which was obviously wrong. I think that somehow a
mosaic >= 1 is glitching the Ycounter for the BG layers to tick one
extra time.

There's a workaround that's not very nice to get everything going
right now. It could very well be that the workaround is hardware
accurate, but I can't help but feel there's a more eloquent way of
doing this.