Update to bsnes v068r04 release.
This should provide hardware-accurate mosaic support in the accurate
renderer, with the exception that I'm still not sure what mid-frame
vertical mosaic or mid-scanline horizontal mosaic writes do. Either
the code I have is correct, or it bypasses the mosaic adjust and gives
the exact H/V positions.

I've also renamed the fast folder to alternative (thinking about
naming it simply alt instead), and started on a brand new PPU
renderer. So far it's just a barebones setup with some MMIO support
and VRAM/OAM/CGRAM writing. I'm not even confident that I can get this
to be faster than the current scanline renderer, but it's the only
avenue that we have left for any kind of significant bsnes speedup, so
I have to try. I'm going to finish up the MMIO stuff first, that way
we have a clean slate with no actual rendering. And then from here we
can try various different approaches.