Update to bsnes v068r10 release.
I've posted a profile-optimized binary of the new PPU.
It is feature-complete, but horizontal mosaic is less accurate. I have
an idea for a mosaic color ring buffer to get it equally accurate, but
I haven't implemented it yet. For now it's just a simple x & ~(mosaic
>> 1) trick that is passable.

Hires blending was left out, as it's more processor intensive and
blargg's NTSC does a better job with that anyway.

There's some OPT vertical positioning issues in the SNES Test
Program's character test; Goodbye, Anthrox has some sort of fast CPU
DMA issue; etc.

Total speedup is a mere 13.5%. Not quite the 50% I wanted in the best
case, but I'll take what I can get.

254->289fps in Zelda 3 on my E8400 now. There's another 15% hiding
with blargg's SMP and 5-10% with blargg's fast DSP, but they lose too
much accuracy. It'd put me at or below Snes9X accuracy, while still
being 50% slower.

SSE2 was performing worse this time, both on x86 and amd64, so I left
that optimization off.

So, barring a miracle, this is about the best it's going to get.