Update to bsnes v068r14 release.
Holy hell, that was a total brain twister. After hours of crazy bit
twiddling and debug printf's, I finally figured out how to allow both
lores and hires scrolling in the accurate PPU renderer. In the
process, I modified the main loop to run from -7 to 255, regardless of
the hires setting, and perform X adjustment inside the tile fetching.
This fixed a strange main/subscreen misalignment issue, so I was able
to restore the proper sub-then-main rendering for the final screen
output stage. Code looks a good bit cleaner this way overall.

I also added load state and save state menus to the tools menu, so you
can use the menubar to load and save to ten slots. I am thinking that
I should nuke the icons. As pretty as they are, it's getting tiresome
trying to find icons for everything, I have no pictures to represent
loading or saving a slot, nor to represent individual slots. I'll just
stick to radios and checkboxes.