Update to bsnes v068r15 release.
This adds a new "Effect Toggle" window to the tools tab list. There's
no menu entry for it yet, oversight, but you can go to another tool
and tab to it.

The effect toggle window lets you toggle background/OAM layers on a
per-priority basis and DSP sound channels. This only works with the
compatibility and performance cores, because I'm not going to allow
accuracy-violating hacks like that into the core.

Non-essential icons have been removed, there's six of them left at the
moment. I am too much of a consistency guy to only have some scattered
around. I know other apps do that, but I'm not going to do that, and I
grow tired of trying to hammer in icons that don't really represent
the actions. Anyway, it still looks pretty good I think.