Update to bsnes v068r17 release.
This hopefully fixes vertical and horizontal mosaic with the accurate
S-PPU core.

It also adds frame-skipping support to the compatibility and
performance S-PPU cores. This is not visible to the user interface,
because I feel that virtually nobody is going to play with frame-
skipping on when they could use a faster emulation with it off.
However, I did use it to enhance the 'Fast Forward' key. When you
press it, it will now turn off video and audio sync, and set a
frameskip of 9. This allows you to speed things up an additional ~60%
over your max theoretical speed without frameskipping. ~35% with the
compatible core, because the PPU is less of a bottleneck for it.
Slowdown was improved, it will turn on audio sync if it is off until
you release the key to ensure you actually get 30fps.

Eventually it'd probably be wise to set a cap on the fast forward
speed, but I hardly think it's an issue on today's computers.

Now if only we can increase that speed by another ten times ...
perhaps Der Langrisser will be tolerable :)