Update to bsnes v068r18 release.
This WIP fixes the Mode7 repeat issue in the accuracy core.

More importantly, it's the first build to include phoenix. There is a
stub GUI that does basically nothing right now. It will give you a
window, a command to close the emulator, and an FPS meter so you can
tell how fast it is. To load a ROM, you have to drag the ROM on top of
the binary. I don't know if it will work if the filename+path has
spaces in it or not, so avoid that to be safe.

The big news is that there's both a 32-bit and 64-bit binary included,
although neither are profiled because I really don't have enough of a
GUI up to profile anyway.

Windows uses: Direct3D + XAudio2 + RawInput - get the latest DX SDK
for XAudio2 to work
Linux uses: OpenGL + ALSA + SDL
You will have to compile yourself to change these driver settings.

For some reason, the 64-bit binary sometimes crashes on start, maybe
1:6 times. So just keep trying. I don't know what's up with that, I'd
appreciate if someone here wanted to debug that for me though :D

One really good bit of news, there was that old hiro bug where
keyboard input would cause the main window to beep. I spied on the
main event loop and, as suspected, the status bar was getting focus
and rejecting key presses. What. The. Fuck. Why would a status bar
ever need focus? So I set WM_DISABLED on it, which luckily leaves the
font color alone. I also had to use WM_DISABLED on the Viewport widget
that I use for video output. These two combined let me have my main
window with no keyboard beeping AND allow tab+shift-tab to work as
you'd expect on other windows, so hooray.

Now, at the moment there's no Manifest included, because Microsoft for
some reason includes the processorArcitecture in the file. So I can't
use the same manifest for 32-bit and 64-bit mode, or the binary will
crash on one or the other. Fuck. So the status bar may look old-school
or something, whatever, it's only temporary.

Next up, my goal is to avoid the hiro icon corruption bullshit by
making phoenix itself try and use an internal resource icon. So just
compile your app with that resource icon and voila, perfect icon. Not
in there yet so you get the white box.

Input is hard-coded, up/down/left/right/z/x/a/s/d/c/apostrophe/return.

Lastly, compilation is ... in a serious state of flux. The code is set
to compile bsnes/phoenix-gtk right now. Try it at your own risk. Give
me a few WIPs to get everything nice and refined. Ubuntu users will
need gcc-4.5, which you can get by adding the Maverick Meerkat
repository, updating apt, installing the gcc-4.5 + g++-4.5 packages,
and then removing and re-updating your apt/sources.list file so you
don't end up fucking your whole system when you run apt again in the
future.

For anyone who can work with all of that, great! Please post a
framerate comparison between 32-bit and 64-bit builds. Any game, any
screen, so long as the FPS is not fluctuating when you measure it (eg
don't do it during an attract sequence.)

If anyone complains about the 64-bit binary not working and it turns
out they are on 32-bit Windows, they are going to be removed from this
WIP forum :P