Update to bsnes v068r21 release.
Has an unprofiled 64-bit Windows binary for some reason. I meant to
make it 32-bit, but forgot to swap the PATH variable to MinGW32, oops.

Changelog:
- adds synchronize video
- adds synchronize audio
- adds mute audio
- adds advanced settings window with driver selection only
- adds the pretty section header thing I started going toward with Qt
- adds a configuration file, saves to bsnes-phoenix.cfg (I don't make
the .bsnes folder yet)
- the status bar shows [A] for accuracy, [C] for compatibility and [P]
for carburetor

I've decided that the phoenix port is going to be strictly a no
bullshit, no bloat port.
No warnings for idiot users, no safeties for things like setting a 0%
gamma or 0% audio output, no bloatey stuff like wide-zoom scaling, no
fullscreen trickery, no 500kb input driver, no movies, no rewind, just
the basics. I want a UI port that I actually enjoy working on. The Qt
port is a nightmare to maintain, and it's not because of bad code but
because of TOO MUCH code.

The Qt port will fill the GUI-equivalent of the performance core,
essentially.