Update to bsnes v068r22 release.
Wanted to torture myself, so I implemented the hardest window of all,
the cheat code editor. I had to sacrifice checkboxes inside lists,
unfortunately, but it's a necessary evil. Maybe some day if we can
port checkboxes inside list items to Windows and GTK+, we can add it
back. It should be the only really apparent GUI sacrifice, though. You
toggle cheats by double-clicking them in the list. Easy to do, but not
apparent.

I also added in the focus policy.