Update to bsnes v068r25 release.
This is basically the v069 release candidate, heavy testing would
really be appreciated. I'd like for v070 to be the first official
phoenix release, but if a major bug slips past us that will ruin it.

The big change is that save state support was added to the PPU-
performance core.