Update to bsnes v069 release.
This release adds some nifty new user interface features, and improves both performance and accuracy. The performance core is now fast enough to maintain a constant 60fps on a first-gen Intel Atom N270 in 98% of games.
Changelog:
* added new effect toggle tool window, which allows toggling of BG/OAM graphics layers and DSP audio channels
* added an option to use the native OS file and folder open dialogs instead of my custom browser
* added a new state selection window
* added frame skipping support, which is only used during fast forwarding; as a result, fast forward is now ~80% faster
* removed unnecessary icons, added workaround for checkbox/radiobox menu icons on Linux/GNOME
* added RTS/CTS support to serial simulation
* all cores: OAM high table even address writes should update OAM latch data register [blargg]
* accuracy core: major improvements to mosaic emulation
* accuracy core: added additional hardware-based caching, resulting in a ~15% speed boost
* accuracy core: emulated CGRAM address invalidation for writes during active display
* performance core: added new S-PPU renderer, resulting in a ~10% speed bost
Also, I'm looking for someone who would be willing to backport bsnes v069 from C++0x to C++98. With this, it should be possible to create a PS3 port of bsnes that runs at full speed with all games. bsnes does not use C++0x very much, so it shouldn't be very hard, a weekend project at best. I just don't have the time to do this myself. Note that I wouldn't be merging these changes, I intend to keep using C++0x, this is needed only because the PS3 SDK uses a really old version of GCC. Please post on the message board if you are interested in helping.