Update to bsnes v070r04 release.
Changelog:
- fixed new config file input driver name (you'll have to delete your
old config, or change to a different driver and back and restart)
- fixed slot loader windows' OK button placement
- fixed nall/directory.hpp when list size was zero
- rewrote nall/function.hpp, no longer requires <functional> or union
tricks
- added state manager

The state manager is a little bit different this time. It's
functionally identical to bsnes/Qt, 100% of the way.
But when you save slots, it stores them in RAM. It only writes the BSA
archive upon ROM unload / program exit.
Yes, this means that technically if the emulator crashes, you'll lose
your states.
But a) that very rarely happens, and b) the old way was thrashing the
disk like crazy, every letter you typed dumped up to 8MB to disk.
With this new method, I can simply store a boolean valid flag before
each slot, and pack the file better.
Before, a save on only slot 3 would be 3*state size (~1.2mb), it will
now be 3bytes+state size (~400kb.)
I have also added a proper signature because of this, so it will
detect when you load an archive for a previous serializer version and
ignore it.
When you go to save (unload the game), if there are no valid slots,
the BSA archive gets unlinked (deleted.)

I am also planning a feature around the now-hidden "slot 0". My idea
is for it to be a fallback slot.
How many times have you loaded a state when you meant to save and
said, "shit, now I lost some of my progress"?
The idea is that whenever you load a state, right before loading, it
will save to slot 0.
When you unload the game, or exit the emulator, it will also save to
slot 0.
You will be able to load from slot 0 from the menu, but not save to
it. It will appear at the bottom of the load list.
And lastly, I'll add an advanced option to auto-load slot 0 if it
exists, which will enable "close the emulator and restart where you
left off." functionality.