Update to bsnes v070r08 release.
Changelog:
- all three ports of phoenix gain the ability to use
ListBox::setCheckable(), checked(row), setChecked(row, checked =
true);
- cheat editor updated to take advantage of this

Some fun differences between the implementations. Windows was the
least flexible, it only lets you have a check at the start of each
item. Luckily that's all I need for my purposes so it'll work. It's
also a lot easier, as now I don't need a ton of extra code to try and
set per-column checkboxes. Now both Windows and Qt can put text into
the first item with the checkbox, but GTK+ cannot. Further, Qt needs
this because even if you hide the checkbox column, it still tries to
search for typed text from the checkbox column. GTK+ does this too,
but unlike GTK+, Qt lacks an API call to set the search column. Since
my code basically has to change this in real-time since you have to
call the setProperty functions after create(), this means I always set
up the checkbox columns regardless of whether or not they are used.
For Qt, I had to work around this and it'll be an annoying edge case
if you try and use setCheckable(true) and then setCheckable(false),
because Qt has no way to clear the checkboxes from an item once you've
enabled them for the first time. But without doing it this way,
there's no way for eg the ROM file loader to allow type-searching, so
that's the way I do it. Windows works the same, and GTK+ has a
separate column (hidden from the phoenix API standpoint) for the
checkboxes, with no column header label text.

All in all, a major hassle, but it was the only really major GUI hit
from leaving Qt, aside from the horror that's going to be the
debugger, which needs all kinds of highly specialized controls.