Update to bsnes v070r12 release.
Changelog:
- **removed support for images with copier headers**
- phoenix/Windows: Label properly refreshes on text changes, fixes
video settings sliders
- alt/ppu-performance: fixed mosaic Voffset bug, fixes Super Bowling
et al
- alt/cpu: fixed CPU::joylatch() reporting, allows serial applications
to work with performance profile
- hooked up SNES::cartridge.basename, allows MSU1 and serial support
with the phoenix UI
- updated UPS patching code for bsnes/Qt, allowing it to compile
again, hidden config option file.bypassPatchCrc32 was removed

To automatically and _recursively_ fix all SNES ROM extensions and
remove all headers, use snespurify:
http://byuu.org/files/snespurify_v01.tar.bz2

Basically, put it in your ROM folder and run it. I strongly suggest
you do this from a command-line so you can see the output. You can
also see what it will do PRIOR to it actually doing it by using this
syntax:
"snespurify /path/to/purify -analyzeonly"

If you stick it in "C:" or "/" and run it, expect it to break your
entire OS. Thanks to the dumbasses that thought it'd be a good idea to
distribute SNES ROMs with .bin as an extension, it will probably end
up renaming your non-SNES .bin files to .sfc, and if any of those are
important, you'll be in trouble.

When the preservation project finally happens, I'll be removing
cartridge header parsing as well and requiring XML. I'll probably have
an internal database for known commercial games, and allow snespurify
to generate XML for the rest.

It also won't extract compressed images yet, nor will it fix
overdumps, and I can probably do a better job ignoring non-SNES .bin
files. It's v01, give it time people.