Update to bsnes v070r16 release.
This mostly contains improvements for nall, like path unification.
This should fix FitzRoy's issue with .. on Windows.

Using /usr/share isn't very portable for Windows with my userpath()
function, I chmod the folder 777 so that it can be deleted by anyone.
Could probably chown it somehow too. Yeah, won't work with su, I
thought that was obvious enough. It uses the user you are currently
logged in as.