Update to bsnes v071 release.
It is very important to note if you choose to use my binaries, the phoenix port does not support ROM images with copier headers, ROM images that do not end in .sfc (for SNES cartridges), .st (for Sufami Turbo cartridges), .bs (for BS-X flash cartridges), or (.gb,.gbc,.sgb) for Game Boy cartridges, it does not support compressed images, it does not support software filters, rewind support, fullscreen mode, IPS patches, cheat searching, movie recording, native OS file dialog boxes, asciiPad simulation, video region cropping, layer/channel toggle, slowdown, etc. It is meant to be minimal, and if this is a problem for you, please stick with the Qt port.
Undoubtedly, no matter how much I emphasize this, people are going to not read this, use the phoenix port, and be put off and not use bsnes anymore. That's very unfortunate, but I am releasing this GUI anyway for those who do share my design goals and aesthetics.
So I will ask a favor of you all, if you see someone mentioning something of the phoenix port not working, please advise them that it is not the general use version, and that they should stick with the Qt port. Thanks in advance.
Changelog:
* fixed a regression in the accuracy/compatibility CPU core with IRQ masking; fixes World Heroes 2
* fixed OAM address reset on $2100 writes for performance PPU core; fixes Mahjongg 2 and others
* DSP-1 always returns high 8-bits of status register; fixes Ace wo Nerae! freeze [Jonas Quinn]
* performance core can now take advantage of serial support
* pixel shaders now use a unified XML format; in the future they will support multi-pass shaders and textures
* major code restructuring
* first public release of phoenix GUI port
* mightymo's cheat code pack is now an external file for the phoenix port
* phoenix port stores cheat codes in XML format as well, unifying all file formats to the same markup language