Update to bsnes v072r07 release.
I'm happy enough with the debugger now. Not 100% up to par with the
old one, but it also does some new things the old one didn't.
- step into / step over are disabled unless they can be done safely
- this means step over is usually grayed unless you hit step into
first, due to bsnes not being opcode-based (you can't skip an opcode
that is half-executed)
- you can now trace console output to disk
- stepping the CPU will print stepped SMP opcodes if the checkbox for
it is on and vice versa
- button added to clear the console log

The only things left are pretty minor:
- add VRAM viewer
- add OAM list viewer
- add CGRAM viewer
- add properties viewer
- add "Proceed" support to CPU/SMP debuggers (glorified breakpoints,
basically)

And then long-term vision stuff to follow that:
- add SA-1 debugger
- add SuperFX debugger
- add "Trace for _n_ opcodes" functionality

But yeah, overall good enough for now. So the next major undertaking
is going to be gutting the bsnes/Qt debugger, which is required to
port that GUI to libsnes.
Once done, bsnes/Qt will be a totally separate program, with a
separate source distribution. I'll merge snesreader and old snesfilter
directly into it. No DLLs.
And after all that is done, I'll spin the entire thing off. Maybe we
can set up an SVN with a list of people who can commit patches to it
and we can host binaries there.