Update to bsnes v072r12 release.
This release defaults DSP-3 loading to use the uPD77C25 core. It also
pre-emptively does the same for the DSP-2 and DSP-4. v072r11 did this
for the DSP-1.

I've also renamed my string<>integer conversion function names:

strhex -> hex
strsigned -> integer
strunsigned -> decimal
strbin -> binary
strdouble -> fp (this one will no doubt be trouble since 'file fp' is
a common idiom. floatingpoint is too long, float and double are
already reserved.)

The reason for this is so that string building is not so hideous.

Pre-C++0x:
    string output = string() << "A: " << strhex<4>(a) << " B: " <<
    strhex<4>(b) << "\n";


Before:
    string output = { "A: ", strhex<4>(a), " B: ", strhex<4>(b), "\n"
    };


After:
    string output = { "A: ", hex<4>(a), " B: ", hex<4>(b), "\n" };