Update to bsnes v073 release.
This release marks a major step forward, offering full low-level emulation of all four DSP coprocessors based on the NEC uPD77C25 processor core. Many people were responsible for this milestone: Dr. Decapitator for the actual decapping and extraction; Lord Nightmare for the cartridges and some special analysis tools; myself, Jonas Quinn and Cydrak for the uPD77C25 emulation; and all of the donors who raised the necessary $1,000 for the necessary hardware and equipment needed to pull this all off. To say thanks to the donors, I am releasing the uPD77C25 emulation core to the public domain, so that everyone can benefit from it.
All four DSP emulations will be improved by this by way of having realistic timing; the DSP-4 will benefit further as the high-level emulation was incomplete and somewhat buggy; and the DSP-3 will benefit the most as the high-levle emulation there was not complete enough to be playable. As a result, most notably, this means bsnes v073 is the first emulator to fully be able to play SD Gundam GX (J)!
As bsnes' primary goal is accuracy, the LLE DSP support renders the old HLE DSP support obsolete. Ergo, I have removed the 166KB of HLE source code, and replaced it with the uPD77C25 core, which comprises a mere 20KB of source code. As this LLE module supports save states, this also means that for the first time, DSP-3 and DSP-4 games have save state support.
On the other hand, this also means that to run any DSP game, you will need the appropriate program ROM. As these are copyrighted, I cannot distribute them nor tell you where to get them. All I can do is provide you with the necessary filenames and hashes.
SHA256:
Code:
	5f2e5ed06b362be023b978b5978813ecb9a07c76592454b45c2a1ed17a0de349  dsp1.bin
	4d42db0f36faef263d6b93f508e8c1c4ae8fc2605fd35e3390ecc02905cd420c  dsp1b.bin
	5efbdf96ed0652790855225964f3e90e6a4d466cfa64df25b110933c6cf94ea1  dsp2.bin
	2e635f72e4d4681148bc35429421c9b946e4f407590e74e31b93b8987b63ba90  dsp3.bin
	63ede17322541c191ed1fdf683872554a0a57306496afc43c59de7c01a6e764a  dsp4.bin
MD5SUM:
Code:
	4865ac61cd758b0f9383fe3d4d3b8694  dsp1.bin
	c8bfb983703a96e1c3d4683105112bc0  dsp1b.bin
	e500ec7f6005e78cb935eea5289c8cc4  dsp2.bin
	c037185c8bbef6313226200dbe5fd07f  dsp3.bin
	fe85065a7023551b0d84941a094435ba  dsp4.bin
All of these files are exactly 8KB (8,192 bytes) in size.
Changelog:
* added NEC uPD77C25 emulation core
* added low-level emulation of the DSP-1, DSP-1B, DSP-2, DSP-3, DSP-4 coprocessors
* removed high-level emulation of the DSP-n coprocessors
* added blargg's libco::ppc.c module, which is far more portable, even running on the PS3
* added software filter support via binary plugins
* added debugger (currently Linux-only); but it is as yet unstable
* added pause shortcut
* updated mightymo's cheat code database

[No archive available]