Update to bsnes v073r04 release.
No uPD96050 emulation here just yet, be patient ;)

Just something I wanted to do first ... I've merged the bgameboy GUI
into the bsnes project, as it's only ~3KB and it means I don't have to
keep manually editing and copying two projects around. You can set ui
:= ui-gameboy to build bgameboy as a standalone, and ui := ui to build
the bsnes you are used to. I will be polishing it to allow changing
the binary output name, not building the SNES portions when not
needed, etc in time.

So that'll end bgameboy standalone releases, at least for now. If it
ever gets important enough to split back out again I can, but I prefer
it this way. bgameboy was never meant to be mainstream anyway, just a
component for bsnes.