Update to bsnes v073r07 release.
I give up for now. This WIP doesn't really improve uPD any, but it
does have major fixes to the debugger.
Debugger is still Linux-only for want of a HexEditor widget, but the
Linux one is now usable.
Before it'd trace and search for breakpoint at sub-1fps. Fixed it to
not break out of core for simple things and it'll run at near-
fullspeed for BPs, half-speed for tracing to file (lots of I/O.)

EDIT: rather obvious, but:
    000f  a = $00da;
    0010  b += drnf;
    0011  rp = a;
    0012  sol = [rp];
    0013  pc.l = sol;


pc.l = sol = datarom[a=00da] is static and unchanging, no matter what
op is fed to DR.
So JP BRCH=0 must be doing something other than just PC=SO.
But even by using all variables and states at that opcode, I can find
no way to make it jump where we want, $1c -> $c1. It's possible it
jumps somewhere else first that then reads from the table to find out
where to jump though. Doesn't look like the $96 it reads from ROM is
right, that code there looks like line noise ... lots of useless,
impossible to hit branch checks.

As for the data ROM itself, $000-0ff (as 16-bit words) is a DIV table.
Formula: value = 32767/offset (thanks Jonas Quinn)

So A=$da, rp=A, ro=32767/$da or 218. 32767/218=150, or $96.
I can't possibly imagine why the hell the game would perform a
constant division here.