Update to bsnes v074 release.
This release adds full low-level emulation of the NEC uPD96050 coprocessor, used by the ST-0010 (F1 Race of Champions II) and the ST-0011 (Hayazashi Nidan Morita Shougi). The former was already playable with HLE, but lacked timing emulation. The latter has never been playable through emulation before now. But as with SD Gundam GX before, you really weren't missing much.
Please see my donations page, and send a small token of thanks to Dr. Decapitator's Paypal address if you don't mind. The NEC uPD96050 proved to be much more difficult than anticipated, and he had to dump the chips multiple times each, spending many days on this.
He is currently working on extraction of the Cx4 (Mega Man X2, Mega Man X3), which initial scans have shown to be an Hitachi HG51B169. Also remaining is the ST-0018, which is used by the only remaining unplayable game: Hayazashi Nidan Morita Shougi 2.
Also new in this release is my own Game Boy emulator. It is being used to provide native Super Game Boy support, built directly into bsnes. This core is released under the GPLv2, but I am willing to grant a more permissive license for other SNES emulators, if anyone is interested.
Of course I cannot compete with the quality of gambatte, and certainly not from only a weeks' worth of work. Currently, there is no Game Boy-side sound output and there are quite a few bugs remaining in its emulation core. I would appreciate any help on this, the Game Boy is not my forte. So yes, we are taking a step back today, so that we may take two steps forward in the future. See my previous posts for more on the advantages my native Game Boy emulation core brings to the project.
As I am in a very good mood, I have posted two pre-compiled binaries this time: the first is a 32-bit Windows binary built with the faster performance profile; and the second is a 64-bit Windows binary built with accuracy profile.
Lastly, the debugger is still Linux-only, but it is now stable enough to be considered usable. Check it out if you like, compile with -DDEBUGGER to enable it.