Update to bsnes v074r07 release.
Fullscreen works on all ports, but Qt is in general pretty wonky at
the moment. Don't really care right now.

F11 toggles fullscreen, there is no menu or status in fullscreen, and
it auto-scales to the highest _even_ multiple that it can.
It copies all other settings (filter, shader, aspect correction, video
region, vsync, async) from your windowed mode settings.
You cannot enable the menu or status bars, because those are a major
pain in the ass to support with both GTK+ and Qt returning nonsense
geometries.
Combine with each platform treating whether or not to handle menus
and/or statuses as part of the geometry or not, and it's a recipe for
pain.
It is pseudo-fullscreen for now, but if we have some spiffy code in
the future to set a true fullscreen that works on Linux (and hopefully
OS X), I'll add a true-FS option to ruby later.
I am not going to hide the mouse cursor at this time, because right
now ruby is in charge of that for mouse capture and it's tricky to
make phoenix also capable of controlling that.
I am not supporting media center users who want to start in
fullscreen, and I am not worrying about idiots who can't figure out
how to exit fullscreen mode.
I am not going out of my way to support multi-monitor setups, Qt had
special handling for it that phoenix does not have.