Update to bsnes v075r02 release.
Changelog:
- added ui-libsnes directory back into source archive; make archive-
all includes it now
- added basic HexEditor widget to phoenix/Windows

I still need a HexEditor class for phoenix/Qt, and the Windows one
lacks a vertical scrollbar. But it does allow the bsnes/Windows
debugger to be built now. You will need to use the goto text box to
type in an address to edit on Windows. The scrollbar really isn't that
useful here anyway, scrolling through 16MB using a 200-pixel-high
scrollbar is pretty terrible.

The bigger problem is that I don't have functions to determine the
width/height of text strings, so the sizes are all screwed up. Since
Windows fonts are so damn big, it is causing lots of ugly wrapping.
And then there's the annoyance that read-only controls on Windows show
up with a grayed out background ... ugh. It looks pretty terrible, so
lots of work will need to go into refining it.